+++ packages/webview_flutter/webview_flutter/example/lib/main.dart
@@ -15,6 +15,7 @@ import 'package:webview_flutter/webview_flutter.dart';
 // #docregion platform_imports
 // Import for Android features.
 import 'package:webview_flutter_android/webview_flutter_android.dart';
+import 'package:webview_flutter_platform_interface/webview_flutter_platform_interface.dart';
 // Import for iOS features.
 import 'package:webview_flutter_wkwebview/webview_flutter_wkwebview.dart';
 // #enddocregion platform_imports
@@ -172,9 +173,9 @@ Page resource error:
           onUrlChange: (UrlChange change) {
             debugPrint('url change to ${change.url}');
           },
-          onHttpBasicAuthRequest: (HttpBasicAuthRequest request) {
+          onHttpAuthRequest: (HttpAuthRequest request) {
             debugPrint(
-                'HTTP basic auth request with host ${request.host} and realm ${request.realm}');
+                'HTTP basic auth request with host ${request.host} and realm ${request.realm ?? '-'}');
             openDialog(request);
           },
         ),
@@ -231,7 +232,7 @@ Page resource error:
     );
   }
 
-  Future<void> openDialog(HttpBasicAuthRequest httpRequest) async {
+  Future<void> openDialog(HttpAuthRequest httpRequest) async {
     final TextEditingController usernameTextController =
         TextEditingController();
     final TextEditingController passwordTextController =
@@ -242,7 +243,7 @@ Page resource error:
       barrierDismissible: false,
       builder: (BuildContext context) {
         return AlertDialog(
-          title: Text('${httpRequest.host}: ${httpRequest.realm}'),
+          title: Text('${httpRequest.host}: ${httpRequest.realm ?? '-'}'),
           content: SingleChildScrollView(
             child: Column(
               mainAxisSize: MainAxisSize.min,
@@ -272,8 +273,10 @@ Page resource error:
             TextButton(
               onPressed: () {
                 httpRequest.onProceed(
-                  usernameTextController.text,
-                  passwordTextController.text,
+                  WebViewCredential(
+                    user: usernameTextController.text,
+                    password: passwordTextController.text,
+                  ),
                 );
                 Navigator.of(context).pop();
               },