+++ packages/webview_flutter/webview_flutter/example/test/main_test.dart
@@ -118,7 +118,7 @@ class FakeNavigationDelegate extends PlatformNavigationDelegate {
   Future<void> setOnUrlChange(UrlChangeCallback onUrlChange) async {}
 
   @override
-  Future<void> setOnHttpBasicAuthRequest(
+  Future<void> setOnHttpAuthRequest(
     HttpAuthRequestCallback handler,
   ) async {}
 }