+++ packages/webview_flutter/webview_flutter_android/example/integration_test/webview_flutter_test.dart
@@ -1140,8 +1140,8 @@ Future<void> main() async {
         PlatformNavigationDelegate(
       const PlatformNavigationDelegateCreationParams(),
     );
-    await navigationDelegate.setOnHttpBasicAuthRequest(
-        (HttpBasicAuthRequest request) => authRequested.complete());
+    await navigationDelegate.setOnHttpAuthRequest(
+        (HttpAuthRequest request) => authRequested.complete());
     await controller.setPlatformNavigationDelegate(navigationDelegate);
 
     // Clear cache so that the auth request is always received and we don't get
@@ -1177,12 +1177,14 @@ Future<void> main() async {
       const PlatformNavigationDelegateCreationParams(),
     );
     await navigationDelegate.setOnPageFinished((_) => pageFinished.complete());
-    await navigationDelegate.setOnHttpBasicAuthRequest(
-      (HttpBasicAuthRequest request) => request.onProceed('user', 'pass'),
+    await navigationDelegate.setOnHttpAuthRequest(
+      (HttpAuthRequest request) => request.onProceed(
+        const WebViewCredential(user: 'user', password: 'password'),
+      ),
     );
     await controller.setPlatformNavigationDelegate(navigationDelegate);
 
-    // Clear cache so that the auth request is always received and we don't get
+    // Clear cache so that the auth request is always received and we do not get
     // a cached response.
     await controller.clearCache();