+++ packages/webview_flutter/webview_flutter_android/test/android_navigation_delegate_test.dart
@@ -474,8 +474,7 @@ void main() {
 
       String? callbackHost;
       String? callbackRealm;
-      androidNavigationDelegate
-          .setOnHttpBasicAuthRequest((HttpBasicAuthRequest request) {
+      androidNavigationDelegate.setOnHttpAuthRequest((HttpAuthRequest request) {
         callbackHost = request.host;
         callbackRealm = request.realm;
       });