+++ packages/webview_flutter/webview_flutter_platform_interface/lib/src/types/types.dart
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-export 'http_basic_auth_request.dart';
+export 'http_auth_request.dart';
 export 'http_response_error.dart';
 export 'javascript_console_message.dart';
 export 'javascript_log_level.dart';
@@ -19,3 +19,4 @@ export 'platform_webview_widget_creation_params.dart';
 export 'url_change.dart';
 export 'web_resource_error.dart';
 export 'webview_cookie.dart';
+export 'webview_credential.dart';